added /api/games route that returns all games in the DB

+++ CalvinAssassinServer/src/main/java/edu/calvin/cs262/Game.java
@@ -107,8 +107,8 @@ public class Game {
                         rs.getString(3),
                         rs.getString(4),
                         rs.getString(5),
-                        rs.getFloat(6),
-                        rs.getFloat(7),
+                        rs.getDouble(6),
+                        rs.getDouble(7),
                         rs.getString(8),
                         rs.getInt(9),
                         rs.getBoolean(10)
